requirements.txt
python-telegram-bot
psycopg2
CLIPSPy
python-dotenv